refactor go experiment files
