+++ go/sigil/experiment.go
@@ -1,20 +1,9 @@
 package sigil
 
 import (
-	"context"
-	"crypto/rand"
-	"crypto/sha1"
-	"encoding/hex"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"maps"
 	"os"
-	"slices"
 	"strconv"
 	"strings"
-	"sync"
-	"time"
 )
 
 const (
@@ -249,1121 +238,4 @@ func TrialRefFromEnv() (*TrialRef, bool) {
 		SuiteVersion: strings.TrimSpace(os.Getenv(EnvSuiteVersion)),
 		TrajectoryID: strings.TrimSpace(os.Getenv(EnvTrajectoryID)),
 	}, true
-}
-
-type ExperimentOptions struct {
-	Client           *Client
-	RunID            string
-	Name             string
-	Suite            *TestSuite
-	Candidate        *Candidate
-	DefaultEvaluator *Evaluator
-	Description      string
-	Tags             []string
-	Metadata         map[string]any
-	AutoFinalize     *bool
-}
-
-type ExperimentRun struct {
-	client           *Client
-	RunID            string
-	Name             string
-	suite            *TestSuite
-	candidate        *Candidate
-	defaultEvaluator Evaluator
-	description      string
-	tags             []string
-	metadata         map[string]any
-	autoFinalize     bool
-
-	mu                   sync.Mutex
-	accepted             int
-	finalized            bool
-	recorders            []*GenerationRecorder
-	trackedIDs           []string
-	activeConversationID string
-	status               string
-}
-
-func NewExperimentRun(opts ExperimentOptions) *ExperimentRun {
-	runID := strings.TrimSpace(opts.RunID)
-	if runID == "" {
-		runID = StableID("exp", opts.Name, experimentRandomHex(8))
-	}
-	name := opts.Name
-	if name == "" {
-		name = runID
-	}
-	defaultEvaluator := Evaluator{EvaluatorID: "sdk", Version: "0", Kind: EvaluatorKindCustom}
-	if opts.DefaultEvaluator != nil {
-		defaultEvaluator = opts.DefaultEvaluator.normalized()
-	}
-	autoFinalize := true
-	if opts.AutoFinalize != nil {
-		autoFinalize = *opts.AutoFinalize
-	}
-	return &ExperimentRun{
-		client:           opts.Client,
-		RunID:            runID,
-		Name:             name,
-		suite:            cloneTestSuite(opts.Suite),
-		candidate:        cloneCandidate(opts.Candidate),
-		defaultEvaluator: defaultEvaluator,
-		description:      opts.Description,
-		tags:             append([]string(nil), opts.Tags...),
-		metadata:         cloneMetadata(opts.Metadata),
-		autoFinalize:     autoFinalize,
-		status:           string(ExperimentStatusRunning),
-	}
-}
-
-func (r *ExperimentRun) Enter(ctx context.Context) error {
-	if r == nil || r.client == nil {
-		return ErrNilClient
-	}
-	metadata := cloneMetadata(r.metadata)
-	if metadata == nil {
-		metadata = map[string]any{}
-	}
-	if r.suite != nil {
-		if r.suite.SuiteID != "" {
-			metadata["suite_id"] = r.suite.SuiteID
-		}
-		if r.suite.Version != "" {
-			metadata["suite_version"] = r.suite.Version
-		}
-	}
-	if r.candidate != nil {
-		maps.Copy(metadata, r.candidate.AsMetadata())
-	}
-	_, err := r.client.CreateExperiment(ctx, CreateExperimentRequest{
-		RunID:       r.RunID,
-		Name:        r.Name,
-		Source:      ExperimentSourceExternal,
-		Description: r.description,
-		Tags:        append([]string(nil), r.tags...),
-		Metadata:    metadata,
-	})
-	return err
-}
-
-func WithExperiment(ctx context.Context, opts ExperimentOptions, fn func(context.Context, *ExperimentRun) error) (run *ExperimentRun, err error) {
-	run = NewExperimentRun(opts)
-	if err := run.Enter(ctx); err != nil {
-		return run, err
-	}
-	defer func() {
-		recovered := recover()
-		if recovered == nil {
-			return
-		}
-		if run.autoFinalize {
-			finalizeCtx, cancel := experimentCleanupContext(ctx)
-			_ = run.Finalize(finalizeCtx, ExperimentStatusFailed, fmt.Sprint(recovered))
-			cancel()
-		}
-		panic(recovered)
-	}()
-	err = fn(run.Context(ctx), run)
-	if !run.autoFinalize {
-		return run, err
-	}
-	finalizeCtx, cancel := experimentCleanupContext(ctx)
-	defer cancel()
-	if err != nil {
-		if finalizeErr := run.Finalize(finalizeCtx, ExperimentStatusFailed, err.Error()); finalizeErr != nil {
-			return run, errors.Join(err, fmt.Errorf("finalize experiment %q: %w", run.RunID, finalizeErr))
-		}
-		return run, err
-	}
-	if finalizeErr := run.Finalize(finalizeCtx, ExperimentStatusCompleted, ""); finalizeErr != nil {
-		return run, finalizeErr
-	}
-	return run, nil
-}
-
-func experimentCleanupContext(ctx context.Context) (context.Context, context.CancelFunc) {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	return context.WithTimeout(context.WithoutCancel(ctx), defaultExperimentRetryTimeout)
-}
-
-func (r *ExperimentRun) Context(ctx context.Context) context.Context {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if r == nil {
-		return ctx
-	}
-	return withExperimentRun(ctx, r)
-}
-
-func (r *ExperimentRun) Suite() *TestSuite {
-	if r == nil {
-		return nil
-	}
-	return cloneTestSuite(r.suite)
-}
-
-func (r *ExperimentRun) Trial(testCase TestCase, opts ...TrialOption) *Trial {
-	testCaseID := strings.TrimSpace(testCase.TestCaseID)
-	testCaseName := firstNonBlank(testCase.Name, testCaseID)
-	return r.trial(testCaseID, testCaseName, opts...)
-}
-
-func (r *ExperimentRun) TrialID(testCaseID string, opts ...TrialOption) *Trial {
-	testCaseID = strings.TrimSpace(testCaseID)
-	testCaseName := testCaseID
-	if r != nil && r.suite != nil {
-		if existing, ok := r.suite.Case(testCaseID); ok {
-			testCaseName = firstNonBlank(existing.Name, testCaseID)
-		}
-	}
-	return r.trial(testCaseID, testCaseName, opts...)
-}
-
-func (r *ExperimentRun) WithTrial(ctx context.Context, testCase TestCase, fn func(context.Context, *Trial) error, opts ...TrialOption) (err error) {
-	if fn == nil {
-		return fmt.Errorf("%w: trial callback is required", ErrExperimentValidationFailed)
-	}
-	ctx = r.Context(ctx)
-	trial := r.Trial(testCase, opts...)
-	if err := trial.Start(ctx); err != nil {
-		return fmt.Errorf("start trial %q: %w", trial.ref.TestCaseID, err)
-	}
-	defer func() {
-		recovered := recover()
-		trialErr := err
-		if recovered != nil {
-			trialErr = fmt.Errorf("trial callback panic: %v", recovered)
-		}
-		if endErr := trial.End(ctx, trialErr); endErr != nil {
-			err = errors.Join(err, fmt.Errorf("end trial %q: %w", trial.ref.TestCaseID, endErr))
-		}
-		if recovered != nil {
-			panic(recovered)
-		}
-	}()
-	return fn(ctx, trial)
-}
-
-func (r *ExperimentRun) WithTrialID(ctx context.Context, testCaseID string, fn func(context.Context, *Trial) error, opts ...TrialOption) (err error) {
-	if fn == nil {
-		return fmt.Errorf("%w: trial callback is required", ErrExperimentValidationFailed)
-	}
-	ctx = r.Context(ctx)
-	trial := r.TrialID(testCaseID, opts...)
-	if err := trial.Start(ctx); err != nil {
-		return fmt.Errorf("start trial %q: %w", trial.ref.TestCaseID, err)
-	}
-	defer func() {
-		recovered := recover()
-		trialErr := err
-		if recovered != nil {
-			trialErr = fmt.Errorf("trial callback panic: %v", recovered)
-		}
-		if endErr := trial.End(ctx, trialErr); endErr != nil {
-			err = errors.Join(err, fmt.Errorf("end trial %q: %w", trial.ref.TestCaseID, endErr))
-		}
-		if recovered != nil {
-			panic(recovered)
-		}
-	}()
-	return fn(ctx, trial)
-}
-
-func (r *ExperimentRun) trial(testCaseID, testCaseName string, opts ...TrialOption) *Trial {
-	if r == nil {
-		return NewTrial(nil, TrialRef{
-			TestCaseID:   strings.TrimSpace(testCaseID),
-			Attempt:      1,
-			TestCaseName: firstNonBlank(testCaseName, strings.TrimSpace(testCaseID)),
-		}, opts...)
-	}
-	ref := TrialRef{
-		RunID:        r.RunID,
-		TestCaseID:   strings.TrimSpace(testCaseID),
-		Attempt:      1,
-		SuiteName:    r.Name,
-		TestCaseName: firstNonBlank(testCaseName, strings.TrimSpace(testCaseID)),
-	}
-	if r.suite != nil {
-		ref.SuiteID = r.suite.SuiteID
-		ref.SuiteVersion = r.suite.Version
-		ref.SuiteName = firstNonBlank(r.suite.Name, r.Name)
-	}
-	t := NewTrial(r.client, ref, opts...)
-	t.experiment = r
-	t.candidate = r.candidate
-	t.defaultEvaluator = r.defaultEvaluator
-	return t
-}
-
-func (r *ExperimentRun) Finalize(ctx context.Context, status ExperimentStatus, errorText string) error {
-	if r == nil || r.client == nil {
-		return ErrNilClient
-	}
-	r.mu.Lock()
-	if r.finalized {
-		r.mu.Unlock()
-		return nil
-	}
-	scoreCount := r.accepted
-	r.status = string(status)
-	r.mu.Unlock()
-	_, err := r.client.FinalizeExperiment(ctx, r.RunID, status, CompleteExperimentOptions{ScoreCount: &scoreCount, Error: errorText})
-	if err != nil {
-		return err
-	}
-	r.mu.Lock()
-	r.finalized = true
-	r.mu.Unlock()
-	return nil
-}
-
-func (r *ExperimentRun) Report(ctx context.Context) (*ExperimentReport, error) {
-	if r == nil || r.client == nil {
-		return nil, ErrNilClient
-	}
-	return r.client.GetExperimentReport(ctx, r.RunID)
-}
-
-func (r *ExperimentRun) URL() string {
-	if r == nil || r.client == nil {
-		return ""
-	}
-	return r.client.ExperimentURL(r.RunID)
-}
-
-func (r *ExperimentRun) AcceptedScores() int {
-	if r == nil {
-		return 0
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return r.accepted
-}
-
-func (r *ExperimentRun) recordAccepted(n int) {
-	r.mu.Lock()
-	r.accepted += n
-	r.mu.Unlock()
-}
-
-func (r *ExperimentRun) TrackGenerationID(generationID string) {
-	if r == nil {
-		return
-	}
-	generationID = strings.TrimSpace(generationID)
-	if generationID == "" {
-		return
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if !slices.Contains(r.trackedIDs, generationID) {
-		r.trackedIDs = append(r.trackedIDs, generationID)
-	}
-}
-
-func (r *ExperimentRun) ResetCapture(conversationID string) string {
-	if r == nil {
-		return ""
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.recorders = nil
-	r.trackedIDs = nil
-	r.activeConversationID = strings.TrimSpace(conversationID)
-	return r.activeConversationID
-}
-
-func (r *ExperimentRun) ProducedGenerationIDs() []string {
-	if r == nil {
-		return nil
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	var ids []string
-	for _, recorder := range r.recorders {
-		if recorder == nil {
-			continue
-		}
-		recorder.mu.Lock()
-		id := recorder.lastGeneration.ID
-		recorder.mu.Unlock()
-		if id != "" && !slices.Contains(ids, id) {
-			ids = append(ids, id)
-		}
-	}
-	for _, id := range r.trackedIDs {
-		if id != "" && !slices.Contains(ids, id) {
-			ids = append(ids, id)
-		}
-	}
-	return ids
-}
-
-func (r *ExperimentRun) ActiveConversationID() string {
-	if r == nil {
-		return ""
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return r.activeConversationID
-}
-
-func (r *ExperimentRun) captureRecorder(recorder *GenerationRecorder) {
-	if r == nil || recorder == nil {
-		return
-	}
-	r.mu.Lock()
-	r.recorders = append(r.recorders, recorder)
-	r.mu.Unlock()
-}
-
-func (r *ExperimentRun) prepareGeneration(start GenerationStart) GenerationStart {
-	r.mu.Lock()
-	conversationID := strings.TrimSpace(start.ConversationID)
-	if conversationID == "" {
-		conversationID = strings.TrimSpace(r.activeConversationID)
-	}
-	if conversationID == "" {
-		conversationID = StableID("conv", r.RunID, experimentRandomHex(8))
-	}
-	r.activeConversationID = conversationID
-	r.mu.Unlock()
-	start.ConversationID = conversationID
-
-	tags := cloneTags(start.Tags)
-	if tags == nil {
-		tags = map[string]string{}
-	}
-	tags[ExperimentRunIDTag] = r.RunID
-	start.Tags = tags
-
-	metadata := cloneMetadata(start.Metadata)
-	if metadata == nil {
-		metadata = map[string]any{}
-	}
-	metadata[ExperimentRunIDMetadataKey] = r.RunID
-	start.Metadata = metadata
-
-	if start.AgentName == "" && r.candidate != nil {
-		start.AgentName = r.candidate.AgentName
-	}
-	if start.AgentVersion == "" && r.candidate != nil {
-		start.AgentVersion = r.candidate.AgentVersion
-	}
-	return start
-}
-
-func prepareGenerationForExperimentRunID(start GenerationStart, runID string) GenerationStart {
-	seed := cloneGenerationStart(start)
-	runID = strings.TrimSpace(runID)
-	if runID == "" {
-		return seed
-	}
-	tags := cloneTags(seed.Tags)
-	if tags == nil {
-		tags = map[string]string{}
-	}
-	tags[ExperimentRunIDTag] = runID
-	seed.Tags = tags
-	metadata := cloneMetadata(seed.Metadata)
-	if metadata == nil {
-		metadata = map[string]any{}
-	}
-	metadata[ExperimentRunIDMetadataKey] = runID
-	seed.Metadata = metadata
-	return seed
-}
-
-type TrialOption func(*Trial)
-
-func WithTrialAttempt(attempt int) TrialOption {
-	return func(t *Trial) {
-		if attempt > 0 {
-			t.ref.Attempt = attempt
-			t.trialID = StableID("trial", t.ref.RunID, t.ref.TestCaseID, t.ref.Attempt)
-			t.generationID = StableID("gen", t.ref.RunID, t.ref.TestCaseID, t.ref.Attempt)
-		}
-	}
-}
-
-func WithTrialMetadata(metadata map[string]any) TrialOption {
-	return func(t *Trial) {
-		maps.Copy(t.metadata, metadata)
-	}
-}
-
-func WithTrialDefaultEvaluator(evaluator Evaluator) TrialOption {
-	return func(t *Trial) {
-		t.defaultEvaluator = evaluator.normalized()
-	}
-}
-
-type Trial struct {
-	client *Client
-	ref    TrialRef
-
-	experiment       *ExperimentRun
-	candidate        *Candidate
-	defaultEvaluator Evaluator
-	metadata         map[string]any
-
-	trialID        string
-	status         TrialStatus
-	conversationID string
-	traceID        string
-	spanID         string
-	errorText      string
-	generationID   string
-
-	generationBound    bool
-	generationExported bool
-	hasGeneration      bool
-	io                 map[string]any
-	trialCreated       bool
-	usage              map[string]any
-	started            time.Time
-
-	buffer      []ScoreItem
-	accepted    int
-	hasFinal    bool
-	finalPassed *bool
-	artifacts   []map[string]any
-}
-
-func NewTrial(client *Client, ref TrialRef, opts ...TrialOption) *Trial {
-	ref.RunID = strings.TrimSpace(ref.RunID)
-	ref.TestCaseID = strings.TrimSpace(ref.TestCaseID)
-	if ref.Attempt <= 0 {
-		ref.Attempt = 1
-	}
-	t := &Trial{
-		client:           client,
-		ref:              ref,
-		defaultEvaluator: Evaluator{EvaluatorID: "sdk", Version: "0", Kind: EvaluatorKindCustom},
-		metadata:         map[string]any{},
-		trialID:          StableID("trial", ref.RunID, ref.TestCaseID, ref.Attempt),
-		status:           TrialStatusRunning,
-		generationID:     StableID("gen", ref.RunID, ref.TestCaseID, ref.Attempt),
-		io:               map[string]any{},
-		usage:            map[string]any{},
-	}
-	for _, opt := range opts {
-		if opt != nil {
-			opt(t)
-		}
-	}
-	return t
-}
-
-func NewTrialFromRef(client *Client, ref *TrialRef, opts ...TrialOption) (*Trial, error) {
-	if ref == nil {
-		return nil, fmt.Errorf("%w: trial ref is required; set SIGIL_EXPERIMENT_ID and SIGIL_TEST_CASE_ID", ErrExperimentValidationFailed)
-	}
-	return NewTrial(client, *ref, opts...), nil
-}
-
-func (t *Trial) Start(ctx context.Context) error {
-	if t == nil || t.client == nil {
-		return ErrNilClient
-	}
-	if t.ref.RunID == "" {
-		return fmt.Errorf("%w: run_id is required", ErrExperimentValidationFailed)
-	}
-	if t.ref.TestCaseID == "" {
-		return fmt.Errorf("%w: test_case_id is required", ErrExperimentValidationFailed)
-	}
-	t.started = time.Now()
-	return t.createTrial(ctx)
-}
-
-func (t *Trial) End(ctx context.Context, err error) error {
-	if t == nil {
-		return ErrNilClient
-	}
-	if err != nil {
-		t.status = TrialStatusErrored
-		t.errorText = err.Error()
-	} else if t.status == TrialStatusRunning {
-		if t.hasFinal {
-			if t.finalPassed != nil && *t.finalPassed {
-				t.status = TrialStatusPassed
-			} else {
-				t.status = TrialStatusFailed
-			}
-		} else {
-			t.status = TrialStatusFailed
-			t.errorText = "trial exited without a final score"
-		}
-	}
-	cleanupCtx, cancel := experimentCleanupContext(ctx)
-	defer cancel()
-	if err := t.createTrial(cleanupCtx); err != nil {
-		return err
-	}
-	_, flushErr := t.Flush(cleanupCtx)
-	if flushErr != nil {
-		t.status = TrialStatusErrored
-		t.errorText = flushErr.Error()
-		if finalizeErr := t.finalizeTrial(cleanupCtx); finalizeErr != nil {
-			return errors.Join(flushErr, fmt.Errorf("finalize trial %q: %w", t.trialID, finalizeErr))
-		}
-		return flushErr
-	}
-	return t.finalizeTrial(cleanupCtx)
-}
-
-func (t *Trial) createTrial(ctx context.Context) error {
-	if t == nil || t.client == nil {
-		return ErrNilClient
-	}
-	if t.trialCreated {
-		return nil
-	}
-	metadata := cloneMetadata(t.metadata)
-	if metadata == nil {
-		metadata = map[string]any{}
-	}
-	if t.ref.TestCaseName != "" {
-		metadata["test_case_name"] = t.ref.TestCaseName
-	}
-	if len(metadata) == 0 {
-		metadata = nil
-	}
-	_, err := t.client.UpsertTrial(ctx, t.ref.RunID, UpsertTrialRequest{
-		TrialID:        t.trialID,
-		TestCaseID:     t.ref.TestCaseID,
-		Attempt:        t.ref.Attempt,
-		Status:         string(TrialStatusRunning),
-		ConversationID: t.conversationID,
-		TraceID:        t.traceID,
-		SpanID:         t.spanID,
-		Metadata:       metadata,
-	})
-	if err != nil {
-		return err
-	}
-	t.trialCreated = true
-	return nil
-}
-
-func (t *Trial) finalizeTrial(ctx context.Context) error {
-	if !t.trialCreated {
-		return nil
-	}
-	backendStatus := "completed"
-	if t.status == TrialStatusErrored {
-		backendStatus = "failed"
-	}
-	req := UpdateTrialRequest{
-		Status:         backendStatus,
-		Error:          t.errorText,
-		ConversationID: t.conversationID,
-		TraceID:        t.traceID,
-	}
-	if v, ok := t.usage["cost"].(float64); ok {
-		req.Cost = &v
-	}
-	if v, ok := t.usage["input_tokens"].(int); ok {
-		req.InputTokens = &v
-	}
-	if v, ok := t.usage["output_tokens"].(int); ok {
-		req.OutputTokens = &v
-	}
-	if !t.started.IsZero() {
-		ms := int(time.Since(t.started).Milliseconds())
-		req.DurationMillis = &ms
-	}
-	_, err := t.client.UpdateTrial(ctx, t.ref.RunID, t.trialID, req)
-	return err
-}
-
-func (t *Trial) BindTrace(traceID, spanID string) *Trial {
-	t.traceID = strings.TrimSpace(traceID)
-	t.spanID = strings.TrimSpace(spanID)
-	return t
-}
-
-func (t *Trial) BindConversation(conversationID string) *Trial {
-	t.conversationID = strings.TrimSpace(conversationID)
-	return t
-}
-
-func (t *Trial) BindGeneration(generationID, conversationID string) *Trial {
-	generationID = strings.TrimSpace(generationID)
-	if generationID != "" {
-		t.generationID = generationID
-		t.generationBound = true
-		t.hasGeneration = true
-	}
-	if strings.TrimSpace(conversationID) != "" {
-		t.conversationID = strings.TrimSpace(conversationID)
-	}
-	return t
-}
-
-type RecordIOOptions struct {
-	Input         any
-	Output        any
-	ModelProvider string
-	ModelName     string
-	AgentName     string
-	InputTokens   *int
-	OutputTokens  *int
-}
-
-func (t *Trial) RecordIO(opts RecordIOOptions) *Trial {
-	if opts.Input != nil {
-		t.io["input_text"] = fmt.Sprint(opts.Input)
-	}
-	if opts.Output != nil {
-		t.io["output_text"] = fmt.Sprint(opts.Output)
-	}
-	if opts.ModelProvider != "" {
-		t.io["model_provider"] = opts.ModelProvider
-	}
-	if opts.ModelName != "" {
-		t.io["model_name"] = opts.ModelName
-	}
-	if opts.AgentName != "" {
-		t.io["agent_name"] = opts.AgentName
-	}
-	if opts.InputTokens != nil {
-		t.io["input_tokens"] = *opts.InputTokens
-		t.usage["input_tokens"] = *opts.InputTokens
-	}
-	if opts.OutputTokens != nil {
-		t.io["output_tokens"] = *opts.OutputTokens
-		t.usage["output_tokens"] = *opts.OutputTokens
-	}
-	if t.hasRecordedGenerationData() {
-		t.hasGeneration = true
-		if t.conversationID == "" {
-			t.conversationID = StableID("conv", t.ref.RunID, t.ref.TestCaseID, t.ref.Attempt)
-		}
-	}
-	return t
-}
-
-func (t *Trial) hasRecordedGenerationData() bool {
-	if t == nil {
-		return false
-	}
-	_, hasInput := t.io["input_text"]
-	_, hasOutput := t.io["output_text"]
-	_, hasInputTokens := t.io["input_tokens"]
-	_, hasOutputTokens := t.io["output_tokens"]
-	return hasInput || hasOutput || hasInputTokens || hasOutputTokens
-}
-
-func (t *Trial) SetUsage(inputTokens, outputTokens *int, cost *float64) *Trial {
-	if inputTokens != nil {
-		t.usage["input_tokens"] = *inputTokens
-	}
-	if outputTokens != nil {
-		t.usage["output_tokens"] = *outputTokens
-	}
-	if cost != nil {
-		t.usage["cost"] = *cost
-	}
-	return t
-}
-
-type ScoreOptions struct {
-	Evaluator            *Evaluator
-	Passed               *bool
-	Explanation          string
-	GenerationID         string
-	GraderConversationID string
-	GraderGenerationID   string
-	GraderTraceID        string
-	Metadata             map[string]any
-}
-
-func (t *Trial) Score(scoreKey string, value ScoreValue, opts ScoreOptions) ScoreItem {
-	if scoreKey == "final" {
-		opts.Passed = inferFinalPassed(value, opts.Passed)
-	}
-	ev := t.defaultEvaluator
-	if opts.Evaluator != nil {
-		ev = opts.Evaluator.normalized()
-	}
-	scoreID := StableID("score", t.ref.RunID, t.trialID, scoreKey, ev.EvaluatorID)
-	metadata := map[string]any{}
-	maps.Copy(metadata, t.metadata)
-	maps.Copy(metadata, opts.Metadata)
-	metadata["task_id"] = t.ref.TestCaseID
-	metadata["trial_id"] = t.trialID
-	metadata["attempt"] = t.ref.Attempt
-	generationID := strings.TrimSpace(opts.GenerationID)
-	if generationID == "" && t.hasGeneration {
-		generationID = t.generationID
-	}
-	item := ScoreItem{
-		ScoreID:              scoreID,
-		EvaluatorID:          ev.EvaluatorID,
-		EvaluatorVersion:     ev.Version,
-		EvaluatorKind:        string(ev.Kind),
-		ScoreKey:             scoreKey,
-		Value:                value,
-		GenerationID:         generationID,
-		TrialID:              t.trialID,
-		ConversationID:       t.conversationID,
-		TraceID:              t.traceID,
-		SpanID:               t.spanID,
-		RunID:                t.ref.RunID,
-		TestCaseID:           t.ref.TestCaseID,
-		GraderConversationID: opts.GraderConversationID,
-		GraderGenerationID:   opts.GraderGenerationID,
-		GraderTraceID:        opts.GraderTraceID,
-		Passed:               opts.Passed,
-		Explanation:          opts.Explanation,
-		Metadata:             metadata,
-		Source:               &ScoreSource{Kind: "experiment", ID: t.ref.RunID},
-	}
-	t.buffer = append(t.buffer, item)
-	if scoreKey == "final" {
-		t.hasFinal = true
-		t.finalPassed = opts.Passed
-	}
-	return item
-}
-
-// FinalScore records the headline score. Boolean scores infer the trial verdict
-// from the value; numeric and string scores require ScoreOptions.Passed.
-func (t *Trial) FinalScore(value ScoreValue, opts ScoreOptions) ScoreItem {
-	opts.Passed = inferFinalPassed(value, opts.Passed)
-	return t.Score("final", value, opts)
-}
-
-func inferFinalPassed(value ScoreValue, passed *bool) *bool {
-	if passed == nil && value.Bool != nil {
-		passed := *value.Bool
-		return &passed
-	}
-	return passed
-}
-
-func (t *Trial) CheckScore(name string, passed bool, opts ScoreOptions) ScoreItem {
-	if opts.Evaluator == nil {
-		ev := Evaluator{
-			EvaluatorID: t.defaultEvaluator.EvaluatorID + "." + name,
-			Version:     t.defaultEvaluator.Version,
-			Kind:        EvaluatorKindDeterministic,
-		}
-		opts.Evaluator = &ev
-	}
-	opts.Passed = &passed
-	return t.Score(name, BoolScoreValue(passed), opts)
-}
-
-func (t *Trial) RubricScore(name string, value ScoreValue, opts ScoreOptions) ScoreItem {
-	if opts.Evaluator == nil {
-		ev := Evaluator{
-			EvaluatorID: t.defaultEvaluator.EvaluatorID + "." + name,
-			Version:     t.defaultEvaluator.Version,
-			Kind:        EvaluatorKindLLMJudge,
-		}
-		opts.Evaluator = &ev
-	}
-	return t.Score(name, value, opts)
-}
-
-type ArtifactOptions struct {
-	Name    string
-	Kind    string
-	MIME    string
-	Content []byte
-	Data    any
-	Text    string
-}
-
-func (t *Trial) Artifact(ctx context.Context, opts ArtifactOptions) (*TrialArtifact, error) {
-	content := opts.Content
-	kind := opts.Kind
-	mime := opts.MIME
-	if len(content) == 0 && opts.Data != nil {
-		raw, err := json.Marshal(opts.Data)
-		if err != nil {
-			return nil, err
-		}
-		content = raw
-		if kind == "" {
-			kind = "json"
-		}
-		if mime == "" {
-			mime = "application/json"
-		}
-	}
-	if len(content) == 0 && opts.Text != "" {
-		content = []byte(opts.Text)
-		if kind == "" {
-			kind = "text"
-		}
-		if mime == "" {
-			mime = "text/plain"
-		}
-	}
-	record, err := t.client.UploadTrialArtifact(ctx, t.ref.RunID, t.trialID, TrialArtifactUpload{
-		Name:    opts.Name,
-		Kind:    kind,
-		MIME:    mime,
-		Content: content,
-	})
-	if err != nil {
-		return nil, err
-	}
-	artifactID := ""
-	if record != nil {
-		artifactID = record.ArtifactID
-	}
-	t.artifacts = append(t.artifacts, map[string]any{"name": opts.Name, "kind": kind, "artifact_id": artifactID})
-	return record, nil
-}
-
-func (t *Trial) Succeed() *Trial {
-	t.status = TrialStatusPassed
-	return t
-}
-
-func (t *Trial) Fail(errorText string) *Trial {
-	t.status = TrialStatusFailed
-	if errorText != "" {
-		t.errorText = errorText
-	}
-	return t
-}
-
-func (t *Trial) ensureGeneration(ctx context.Context) error {
-	if t.generationExported || !t.hasRecordedGenerationData() {
-		return nil
-	}
-	if t.generationBound && t.client.hasRecordedGenerationID(t.generationID) {
-		if err := t.client.Flush(ctx); err != nil {
-			return err
-		}
-		t.generationExported = true
-		return nil
-	}
-	caseInput := ""
-	if t.experiment != nil && t.experiment.suite != nil {
-		if tc, ok := t.experiment.suite.Case(t.ref.TestCaseID); ok && tc.Input != nil {
-			caseInput = fmt.Sprint(tc.Input)
-		}
-	}
-	provider := firstNonBlank(firstString(t.io["model_provider"]), candidateModelProvider(t.candidate), "eval")
-	model := firstNonBlank(firstString(t.io["model_name"]), candidateModelName(t.candidate), "experiment")
-	agentName := firstNonBlank(firstString(t.io["agent_name"]), candidateAgentName(t.candidate))
-	ctx, recorder := t.client.StartGeneration(ctx, GenerationStart{
-		ID:             t.generationID,
-		ConversationID: t.conversationID,
-		Model:          ModelRef{Provider: provider, Name: model},
-		AgentName:      agentName,
-		OperationName:  "invoke_agent",
-		Tags:           map[string]string{"experiment.run_id": t.ref.RunID, "task_id": t.ref.TestCaseID},
-		Metadata: map[string]any{
-			"experiment_run_id": t.ref.RunID,
-			"task_id":           t.ref.TestCaseID,
-			"trial_id":          t.trialID,
-			"attempt":           t.ref.Attempt,
-		},
-	})
-	usage := TokenUsage{}
-	if v, ok := t.io["input_tokens"].(int); ok {
-		usage.InputTokens = int64(v)
-	}
-	if v, ok := t.io["output_tokens"].(int); ok {
-		usage.OutputTokens = int64(v)
-	}
-	recorder.SetResult(Generation{
-		ID:             t.generationID,
-		ConversationID: t.conversationID,
-		Model:          ModelRef{Provider: provider, Name: model},
-		AgentName:      agentName,
-		Input:          textMessages(RoleUser, firstNonBlank(firstString(t.io["input_text"]), caseInput)),
-		Output:         textMessages(RoleAssistant, firstString(t.io["output_text"])),
-		Usage:          usage,
-	}, nil)
-	recorder.End()
-	if err := recorder.Err(); err != nil {
-		return err
-	}
-	if err := t.client.Flush(ctx); err != nil {
-		return err
-	}
-	t.generationExported = true
-	return nil
-}
-
-func (t *Trial) Flush(ctx context.Context) (int, error) {
-	if t == nil || t.client == nil {
-		return 0, ErrNilClient
-	}
-	if len(t.buffer) == 0 {
-		return 0, nil
-	}
-	if err := t.ensureGeneration(ctx); err != nil {
-		return 0, err
-	}
-	if err := t.client.Flush(ctx); err != nil {
-		return 0, err
-	}
-	pending := append([]ScoreItem(nil), t.buffer...)
-	response, err := t.client.ExportScores(ctx, pending)
-	if err != nil {
-		return 0, err
-	}
-	accepted, err := acceptedOrError(response)
-	if err != nil {
-		return 0, err
-	}
-	t.buffer = t.buffer[len(pending):]
-	t.accepted += accepted
-	if t.experiment != nil {
-		t.experiment.recordAccepted(accepted)
-	}
-	return accepted, nil
-}
-
-func (t *Trial) AcceptedScores() int {
-	if t == nil {
-		return 0
-	}
-	return t.accepted
-}
-
-func StableID(prefix string, parts ...any) string {
-	values := make([]string, len(parts))
-	for i, part := range parts {
-		if part != nil {
-			values[i] = fmt.Sprint(part)
-		}
-	}
-	digest := sha1.Sum([]byte(strings.Join(values, "\x1f")))
-	return prefix + "-" + hex.EncodeToString(digest[:])[:16]
-}
-
-func acceptedOrError(response *ExportScoresResponse) (int, error) {
-	if response == nil {
-		return 0, fmt.Errorf("%w: empty response", ErrScoreExportFailed)
-	}
-	rejected := response.Rejected()
-	if len(rejected) == 0 {
-		return response.AcceptedCount(), nil
-	}
-	parts := make([]string, len(rejected))
-	for i, result := range rejected {
-		detail := result.Error
-		if detail == "" {
-			detail = "rejected"
-		}
-		parts[i] = result.ScoreID + ": " + detail
-	}
-	return 0, fmt.Errorf("%w: rejected %d score(s): %s", ErrScoreExportFailed, len(rejected), strings.Join(parts, "; "))
-}
-
-func cloneTestSuite(in *TestSuite) *TestSuite {
-	if in == nil {
-		return nil
-	}
-	out := *in
-	out.Tags = append([]string(nil), in.Tags...)
-	out.TestCases = cloneTestCases(in.TestCases)
-	return &out
-}
-
-func cloneTestCases(in []TestCase) []TestCase {
-	if len(in) == 0 {
-		return nil
-	}
-	out := make([]TestCase, len(in))
-	for i := range in {
-		out[i] = cloneTestCase(in[i])
-	}
-	return out
-}
-
-func cloneTestCase(in TestCase) TestCase {
-	in.Tags = append([]string(nil), in.Tags...)
-	in.Metadata = cloneMetadata(in.Metadata)
-	return in
-}
-
-func cloneCandidate(in *Candidate) *Candidate {
-	if in == nil {
-		return nil
-	}
-	out := *in
-	return &out
-}
-
-func experimentRandomHex(n int) string {
-	if n <= 0 {
-		return ""
-	}
-	buf := make([]byte, n)
-	if _, err := rand.Read(buf); err != nil {
-		return fmt.Sprintf("%d", time.Now().UnixNano())
-	}
-	return hex.EncodeToString(buf)
-}
-
-func firstNonBlank(values ...string) string {
-	for _, value := range values {
-		if strings.TrimSpace(value) != "" {
-			return strings.TrimSpace(value)
-		}
-	}
-	return ""
-}
-
-func firstString(values ...any) string {
-	for _, value := range values {
-		switch typed := value.(type) {
-		case string:
-			if typed != "" {
-				return typed
-			}
-		case fmt.Stringer:
-			if typed.String() != "" {
-				return typed.String()
-			}
-		}
-	}
-	return ""
-}
-
-func textMessages(role Role, text string) []Message {
-	if text == "" {
-		return nil
-	}
-	if role == RoleAssistant {
-		return []Message{AssistantTextMessage(text)}
-	}
-	return []Message{UserTextMessage(text)}
-}
-
-func candidateModelProvider(c *Candidate) string {
-	if c == nil {
-		return ""
-	}
-	return c.ModelProvider
-}
-
-func candidateModelName(c *Candidate) string {
-	if c == nil {
-		return ""
-	}
-	return c.ModelName
-}
-
-func candidateAgentName(c *Candidate) string {
-	if c == nil {
-		return ""
-	}
-	return c.AgentName
 }
