+++ go/sigil/experiments_test.go
@@ -11,7 +11,6 @@ import (
 	"testing"
 	"time"
 
-	sigilv1 "github.com/grafana/sigil-sdk/go/proto/sigil/v1"
 	"go.opentelemetry.io/otel/trace/noop"
 )
 
@@ -480,767 +479,5 @@ func TestListExperimentScoresParsesTypedScores(t *testing.T) {
 	score := response.Items[0]
 	if score.ScoreID != "score-1" || score.ScoreType != ScoreTypeNumber || score.Value.Number == nil || *score.Value.Number != 0.75 {
 		t.Fatalf("unexpected typed score: %#v", score)
-	}
-}
-
-func TestTrialLifecycleCreatesTypedTrialAndFinalScore(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-	recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	suite := &TestSuite{SuiteID: "smoke", Name: "Smoke", Version: "1.2.0", TestCases: []TestCase{{TestCaseID: "add", Name: "Addition", Input: "2+2"}}}
-	exp := NewExperimentRun(ExperimentOptions{Client: client, RunID: "run-1", Name: "smoke run", Suite: suite})
-	trial := exp.Trial(suite.TestCases[0], WithTrialMetadata(map[string]any{
-		"task_category": "math",
-		"task_id":       "caller-task",
-		"trial_id":      "caller-trial",
-		"attempt":       99,
-	}))
-	if err := trial.Start(context.Background()); err != nil {
-		t.Fatalf("start trial: %v", err)
-	}
-	verifier := Evaluator{EvaluatorID: "exact", Version: "1", Kind: EvaluatorKindDeterministic}
-	trial.FinalScore(NumberScoreValue(1), ScoreOptions{
-		Passed:      BoolPtr(true),
-		Explanation: "matched",
-		Evaluator:   &verifier,
-		Metadata: map[string]any{
-			"trial_id": "score-option-trial",
-			"attempt":  42,
-		},
-	})
-	if err := trial.End(context.Background(), nil); err != nil {
-		t.Fatalf("end trial: %v", err)
-	}
-
-	createReq := recorder.request(0)
-	if createReq.Method != http.MethodPost || createReq.Path != "/api/v1/experiment-runs/run-1/trials" {
-		t.Fatalf("unexpected trial create: %#v", createReq)
-	}
-	createMetadata := createReq.Payload["metadata"].(map[string]any)
-	if createMetadata["task_category"] != "math" || createMetadata["task_id"] != "caller-task" || createMetadata["test_case_name"] != "Addition" {
-		t.Fatalf("trial metadata omitted from upsert: %#v", createMetadata)
-	}
-	scoreReq := recorder.request(1)
-	score := scoreReq.Payload["scores"].([]any)[0].(map[string]any)
-	if score["experiment_id"] != "run-1" || score["test_case_id"] != "add" || score["trial_id"] == "" {
-		t.Fatalf("unexpected score: %#v", score)
-	}
-	scoreMetadata := score["metadata"].(map[string]any)
-	if scoreMetadata["task_id"] != "add" || scoreMetadata["trial_id"] != trial.trialID || scoreMetadata["attempt"] != float64(1) {
-		t.Fatalf("score metadata must keep SDK identifiers authoritative: %#v", scoreMetadata)
-	}
-	if scoreMetadata["task_category"] != "math" {
-		t.Fatalf("score metadata omitted caller metadata: %#v", scoreMetadata)
-	}
-	if _, ok := score["generation_id"]; ok {
-		t.Fatalf("score without RecordIO/BindGeneration must not send generation_id: %#v", score)
-	}
-	updateReq := recorder.request(2)
-	if updateReq.Method != http.MethodPatch || updateReq.Payload["status"] != "completed" {
-		t.Fatalf("unexpected trial update: %#v", updateReq)
-	}
-}
-
-func TestTrialEndCreatesTrialWhenStartWasSkipped(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-no-start"})
-	recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-no-start"})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	trial := NewTrial(client, TrialRef{RunID: "run-no-start", TestCaseID: "case-no-start"})
-	trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-	if err := trial.End(context.Background(), nil); err != nil {
-		t.Fatalf("end trial without start: %v", err)
-	}
-	if recorder.requestCount() != 3 {
-		t.Fatalf("expected trial create, score export, and update requests, got %d", recorder.requestCount())
-	}
-	if req := recorder.request(0); req.Method != http.MethodPost || req.Path != "/api/v1/experiment-runs/run-no-start/trials" {
-		t.Fatalf("unexpected trial create request: %#v", req)
-	}
-	if req := recorder.request(2); req.Method != http.MethodPatch || req.Path != "/api/v1/experiment-runs/run-no-start/trials/"+trial.trialID {
-		t.Fatalf("unexpected trial update request: %#v", req)
-	}
-}
-
-func TestRecordIOTokensAreIncludedInTrialUpdate(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-usage"})
-	recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-usage"})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	inputTokens := 11
-	outputTokens := 7
-	trial := NewTrial(client, TrialRef{RunID: "run-usage", TestCaseID: "case-usage"})
-	if err := trial.Start(context.Background()); err != nil {
-		t.Fatalf("start trial: %v", err)
-	}
-	trial.RecordIO(RecordIOOptions{Input: "question", Output: "answer", InputTokens: &inputTokens, OutputTokens: &outputTokens})
-	trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-	if err := trial.End(context.Background(), nil); err != nil {
-		t.Fatalf("end trial: %v", err)
-	}
-	updateReq := recorder.request(2)
-	if updateReq.Payload["input_tokens"] != float64(inputTokens) || updateReq.Payload["output_tokens"] != float64(outputTokens) {
-		t.Fatalf("expected token usage on trial update, got %#v", updateReq.Payload)
-	}
-}
-
-func TestTrialEndUsesCleanupContextAfterCallerContextCanceled(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-canceled"})
-	recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-canceled"})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	trial := NewTrial(client, TrialRef{RunID: "run-canceled", TestCaseID: "case-canceled"})
-	if err := trial.Start(context.Background()); err != nil {
-		t.Fatalf("start trial: %v", err)
-	}
-	trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-	if err := trial.End(ctx, nil); err != nil {
-		t.Fatalf("end trial with canceled context: %v", err)
-	}
-	if recorder.requestCount() != 3 {
-		t.Fatalf("expected create, score export, and update requests, got %d", recorder.requestCount())
-	}
-	if req := recorder.request(1); req.Method != http.MethodPost || req.Path != "/api/v1/scores:export" {
-		t.Fatalf("unexpected score export request: %#v", req)
-	}
-	if req := recorder.request(2); req.Method != http.MethodPatch || req.Path != "/api/v1/experiment-runs/run-canceled/trials/"+trial.trialID {
-		t.Fatalf("unexpected trial update request: %#v", req)
-	}
-}
-
-func TestTrialEndFinalizesFailedWhenScoreFlushFails(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-flush-fails"})
-	recorder.push(http.StatusInternalServerError, map[string]any{"error": "score export failed"})
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-flush-fails"})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	client.config.GenerationExport.MaxRetries = 0
-	trial := NewTrial(client, TrialRef{RunID: "run-flush-fails", TestCaseID: "case-flush-fails"})
-	if err := trial.Start(context.Background()); err != nil {
-		t.Fatalf("start trial: %v", err)
-	}
-	trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-
-	if err := trial.End(context.Background(), nil); err == nil {
-		t.Fatal("expected score flush error")
-	}
-	if recorder.requestCount() != 3 {
-		t.Fatalf("expected trial create, score export, and failed update, got %d requests", recorder.requestCount())
-	}
-	if req := recorder.request(1); req.Method != http.MethodPost || req.Path != "/api/v1/scores:export" {
-		t.Fatalf("unexpected score export request: %#v", req)
-	}
-	updateReq := recorder.request(2)
-	if updateReq.Method != http.MethodPatch || updateReq.Path != "/api/v1/experiment-runs/run-flush-fails/trials/"+trial.trialID {
-		t.Fatalf("unexpected trial update request: %#v", updateReq)
-	}
-	if updateReq.Payload["status"] != "failed" || updateReq.Payload["error"] == "" {
-		t.Fatalf("expected failed trial update with error, got %#v", updateReq.Payload)
-	}
-}
-
-func TestExperimentRunTrialIDWithoutSuiteDoesNotPanic(t *testing.T) {
-	exp := NewExperimentRun(ExperimentOptions{RunID: "run-no-suite", Name: "no suite"})
-	trial := exp.TrialID("case-1")
-	if trial == nil {
-		t.Fatal("expected trial")
-	}
-	if trial.ref.TestCaseID != "case-1" || trial.ref.TestCaseName != "case-1" {
-		t.Fatalf("unexpected trial ref: %#v", trial.ref)
-	}
-}
-
-func TestExperimentRunCopiesSuiteAtBoundary(t *testing.T) {
-	suite := &TestSuite{
-		SuiteID: "suite-1",
-		TestCases: []TestCase{{
-			TestCaseID: "case-1",
-			Tags:       []string{"original"},
-			Metadata:   map[string]any{"key": "value"},
-		}},
-	}
-	exp := NewExperimentRun(ExperimentOptions{RunID: "run-1", Suite: suite})
-
-	suite.TestCases[0].TestCaseID = "mutated"
-	suite.TestCases[0].Tags[0] = "mutated"
-	suite.TestCases[0].Metadata["key"] = "mutated"
-
-	cases := exp.Suite().Cases()
-	if len(cases) != 1 || cases[0].TestCaseID != "case-1" || cases[0].Tags[0] != "original" || cases[0].Metadata["key"] != "value" {
-		t.Fatalf("experiment retained caller-owned suite data: %#v", cases)
-	}
-
-	cases[0].Tags[0] = "caller-mutated"
-	cases[0].Metadata["key"] = "caller-mutated"
-	fresh := exp.Suite().Cases()
-	if fresh[0].Tags[0] != "original" || fresh[0].Metadata["key"] != "value" {
-		t.Fatalf("suite accessor returned mutable internals: %#v", fresh)
-	}
-}
-
-func TestExperimentRunWithTrialEndsLifecycle(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-	recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-	recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	testCase := TestCase{TestCaseID: "case-1", Name: "Case 1", Metadata: map[string]any{"task": "demo"}}
-	exp := NewExperimentRun(ExperimentOptions{Client: client, RunID: "run-1", Name: "run"})
-	evaluator := Evaluator{EvaluatorID: "exact", Version: "1", Kind: EvaluatorKindDeterministic}
-	err := exp.WithTrial(context.Background(), testCase, func(ctx context.Context, trial *Trial) error {
-		run, ok := experimentRunFromContext(ctx)
-		if !ok || run.RunID != "run-1" {
-			t.Fatalf("expected experiment-aware callback context, got run=%#v ok=%v", run, ok)
-		}
-		trial.FinalScore(BoolScoreValue(true), ScoreOptions{Evaluator: &evaluator})
-		return nil
-	}, WithTrialMetadata(testCase.Metadata))
-	if err != nil {
-		t.Fatalf("with trial: %v", err)
-	}
-	if recorder.requestCount() != 3 {
-		t.Fatalf("expected create, score export, update; got %d request(s)", recorder.requestCount())
-	}
-	if got := exp.AcceptedScores(); got != 1 {
-		t.Fatalf("expected one accepted score, got %d", got)
-	}
-	updateReq := recorder.request(2)
-	if updateReq.Payload["status"] != "completed" {
-		t.Fatalf("expected completed trial update, got %#v", updateReq.Payload)
-	}
-}
-
-func TestExperimentRunWithTrialFinalizesFailedOnPanic(t *testing.T) {
-	type trialRunner func(*ExperimentRun, context.Context, TestCase, func(context.Context, *Trial) error) error
-	tests := []struct {
-		name string
-		run  trialRunner
-	}{
-		{
-			name: "WithTrial",
-			run: func(exp *ExperimentRun, ctx context.Context, testCase TestCase, fn func(context.Context, *Trial) error) error {
-				return exp.WithTrial(ctx, testCase, fn)
-			},
-		},
-		{
-			name: "WithTrialID",
-			run: func(exp *ExperimentRun, ctx context.Context, testCase TestCase, fn func(context.Context, *Trial) error) error {
-				return exp.WithTrialID(ctx, testCase.TestCaseID, fn)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			recorder := &experimentRecorder{}
-			recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-			recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-			recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-			server := httptest.NewServer(recorder.handler(t))
-			defer server.Close()
-
-			client := newExperimentTestClient(t, server.URL)
-			exp := NewExperimentRun(ExperimentOptions{Client: client, RunID: "run-1", Name: "run"})
-			testCase := TestCase{TestCaseID: "case-1", Name: "Case 1"}
-			evaluator := Evaluator{EvaluatorID: "exact", Version: "1", Kind: EvaluatorKindDeterministic}
-
-			var recovered any
-			func() {
-				defer func() {
-					recovered = recover()
-				}()
-				_ = tt.run(exp, context.Background(), testCase, func(_ context.Context, trial *Trial) error {
-					trial.Succeed()
-					trial.FinalScore(BoolScoreValue(true), ScoreOptions{Evaluator: &evaluator})
-					panic("boom")
-				})
-			}()
-
-			if recovered != "boom" {
-				t.Fatalf("expected panic to be rethrown, got %#v", recovered)
-			}
-			if recorder.requestCount() != 3 {
-				t.Fatalf("expected create, score export, update; got %d request(s)", recorder.requestCount())
-			}
-			updateReq := recorder.request(2)
-			if updateReq.Payload["status"] != "failed" {
-				t.Fatalf("expected failed trial update, got %#v", updateReq.Payload)
-			}
-			if updateReq.Payload["error"] != "trial callback panic: boom" {
-				t.Fatalf("expected panic error in trial update, got %#v", updateReq.Payload)
-			}
-		})
-	}
-}
-
-func TestExperimentRunWithTrialErrorOverridesTerminalStatus(t *testing.T) {
-	type trialRunner func(*ExperimentRun, context.Context, TestCase, func(context.Context, *Trial) error) error
-	tests := []struct {
-		name string
-		run  trialRunner
-	}{
-		{
-			name: "WithTrial",
-			run: func(exp *ExperimentRun, ctx context.Context, testCase TestCase, fn func(context.Context, *Trial) error) error {
-				return exp.WithTrial(ctx, testCase, fn)
-			},
-		},
-		{
-			name: "WithTrialID",
-			run: func(exp *ExperimentRun, ctx context.Context, testCase TestCase, fn func(context.Context, *Trial) error) error {
-				return exp.WithTrialID(ctx, testCase.TestCaseID, fn)
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			recorder := &experimentRecorder{}
-			recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-			recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-			recorder.push(http.StatusOK, map[string]any{"trial_id": "trial-1"})
-			server := httptest.NewServer(recorder.handler(t))
-			defer server.Close()
-
-			client := newExperimentTestClient(t, server.URL)
-			exp := NewExperimentRun(ExperimentOptions{Client: client, RunID: "run-1", Name: "run"})
-			testCase := TestCase{TestCaseID: "case-1", Name: "Case 1"}
-			evaluator := Evaluator{EvaluatorID: "exact", Version: "1", Kind: EvaluatorKindDeterministic}
-
-			err := tt.run(exp, context.Background(), testCase, func(_ context.Context, trial *Trial) error {
-				trial.Succeed()
-				trial.FinalScore(BoolScoreValue(true), ScoreOptions{Evaluator: &evaluator})
-				return errors.New("callback failed")
-			})
-			if err == nil || err.Error() != "callback failed" {
-				t.Fatalf("expected callback error, got %v", err)
-			}
-			if recorder.requestCount() != 3 {
-				t.Fatalf("expected create, score export, update; got %d request(s)", recorder.requestCount())
-			}
-			updateReq := recorder.request(2)
-			if updateReq.Payload["status"] != "failed" || updateReq.Payload["error"] != "callback failed" {
-				t.Fatalf("expected failed trial update from callback error, got %#v", updateReq.Payload)
-			}
-		})
-	}
-}
-
-func TestExperimentRunWithTrialRequiresTestCaseID(t *testing.T) {
-	client := newExperimentTestClient(t, "http://example.invalid")
-	exp := NewExperimentRun(ExperimentOptions{Client: client, RunID: "run-1", Name: "run"})
-	err := exp.WithTrial(context.Background(), TestCase{}, func(context.Context, *Trial) error {
-		return nil
-	})
-	if !errors.Is(err, ErrExperimentValidationFailed) {
-		t.Fatalf("expected validation error, got %v", err)
-	}
-}
-
-func TestFinalScoreBoolInfersPassed(t *testing.T) {
-	trial := NewTrial(nil, TrialRef{RunID: "run-1", TestCaseID: "case-1"})
-	score := trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-	if score.Passed == nil || !*score.Passed {
-		t.Fatalf("expected bool final score to infer passed, got %#v", score.Passed)
-	}
-	if trial.finalPassed == nil || !*trial.finalPassed {
-		t.Fatalf("expected trial final verdict to be true, got %#v", trial.finalPassed)
-	}
-
-	trial = NewTrial(nil, TrialRef{RunID: "run-1", TestCaseID: "case-2"})
-	score = trial.Score("final", BoolScoreValue(false), ScoreOptions{})
-	if score.Passed == nil || *score.Passed {
-		t.Fatalf("expected Score(\"final\", bool) to infer failed, got %#v", score.Passed)
-	}
-	if trial.finalPassed == nil || *trial.finalPassed {
-		t.Fatalf("expected trial final verdict to be false, got %#v", trial.finalPassed)
-	}
-}
-
-func TestFinalScoreNumberRequiresExplicitPassed(t *testing.T) {
-	trial := NewTrial(nil, TrialRef{RunID: "run-1", TestCaseID: "case-1"})
-	score := trial.FinalScore(NumberScoreValue(0.2), ScoreOptions{})
-	if score.Passed != nil {
-		t.Fatalf("expected numeric final score to require explicit passed, got %#v", score.Passed)
-	}
-	if trial.finalPassed != nil {
-		t.Fatalf("expected trial final verdict to remain nil, got %#v", trial.finalPassed)
-	}
-}
-
-func TestWithExperimentFinalizesFailedOnPanic(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"run": experimentBody(map[string]any{"experiment_id": "run-panic"})})
-	recorder.push(http.StatusOK, map[string]any{"run": experimentBody(map[string]any{"experiment_id": "run-panic", "status": "failed"})})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	defer func() {
-		recovered := recover()
-		if recovered != "boom" {
-			t.Fatalf("expected panic to be rethrown, got %#v", recovered)
-		}
-		if recorder.requestCount() != 2 {
-			t.Fatalf("expected create and failed finalize requests, got %d", recorder.requestCount())
-		}
-		finalizeReq := recorder.request(1)
-		if finalizeReq.Method != http.MethodPost || finalizeReq.Path != "/api/v1/experiment-runs/run-panic:finalize" {
-			t.Fatalf("unexpected finalize request: %#v", finalizeReq)
-		}
-		if finalizeReq.Payload["status"] != "failed" || finalizeReq.Payload["error"] != "boom" {
-			t.Fatalf("expected failed finalize payload, got %#v", finalizeReq.Payload)
-		}
-	}()
-
-	_, _ = WithExperiment(context.Background(), ExperimentOptions{Client: client, RunID: "run-panic", Name: "panic"}, func(context.Context, *ExperimentRun) error {
-		panic("boom")
-	})
-}
-
-func TestWithExperimentFinalizesWithCleanupContextAfterCancel(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusOK, map[string]any{"run": experimentBody(map[string]any{"experiment_id": "run-canceled"})})
-	recorder.push(http.StatusOK, map[string]any{"run": experimentBody(map[string]any{"experiment_id": "run-canceled", "status": "failed"})})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	ctx, cancel := context.WithCancel(context.Background())
-	_, err := WithExperiment(ctx, ExperimentOptions{Client: client, RunID: "run-canceled", Name: "canceled"}, func(context.Context, *ExperimentRun) error {
-		cancel()
-		return context.Canceled
-	})
-	if !errors.Is(err, context.Canceled) {
-		t.Fatalf("expected context.Canceled, got %v", err)
-	}
-	if recorder.requestCount() != 2 {
-		t.Fatalf("expected create and failed finalize requests, got %d", recorder.requestCount())
-	}
-	finalizeReq := recorder.request(1)
-	if finalizeReq.Method != http.MethodPost || finalizeReq.Path != "/api/v1/experiment-runs/run-canceled:finalize" {
-		t.Fatalf("unexpected finalize request: %#v", finalizeReq)
-	}
-	if finalizeReq.Payload["status"] != "failed" || finalizeReq.Payload["error"] != context.Canceled.Error() {
-		t.Fatalf("expected failed finalize payload, got %#v", finalizeReq.Payload)
-	}
-}
-
-func TestTrialRefEnvRoundTrip(t *testing.T) {
-	ref := TrialRef{RunID: "run-4", TestCaseID: "c1", Attempt: 3, SuiteID: "s", SuiteVersion: "2.0.0"}
-	env := ref.ToEnv()
-	if env[EnvExperimentID] != "run-4" || env[EnvAttempt] != "3" {
-		t.Fatalf("unexpected env: %#v", env)
-	}
-	t.Setenv(EnvExperimentID, env[EnvExperimentID])
-	t.Setenv(EnvTestCaseID, env[EnvTestCaseID])
-	t.Setenv(EnvAttempt, env[EnvAttempt])
-	t.Setenv(EnvSuiteID, env[EnvSuiteID])
-	t.Setenv(EnvSuiteVersion, env[EnvSuiteVersion])
-	restored, ok := TrialRefFromEnv()
-	if !ok || restored.RunID != "run-4" || restored.TestCaseID != "c1" || restored.Attempt != 3 {
-		t.Fatalf("unexpected ref: %#v ok=%v", restored, ok)
-	}
-}
-
-type capturingExperimentExporter struct {
-	mu       sync.Mutex
-	requests []*sigilv1.ExportGenerationsRequest
-}
-
-func (e *capturingExperimentExporter) Export(_ context.Context, request *sigilv1.ExportGenerationsRequest) (*sigilv1.ExportGenerationsResponse, error) {
-	e.mu.Lock()
-	e.requests = append(e.requests, request)
-	e.mu.Unlock()
-	response := &sigilv1.ExportGenerationsResponse{}
-	for _, generation := range request.GetGenerations() {
-		response.Results = append(response.Results, &sigilv1.ExportGenerationResult{
-			GenerationId: generation.GetId(),
-			Accepted:     true,
-		})
-	}
-	return response, nil
-}
-
-func (e *capturingExperimentExporter) Shutdown(context.Context) error { return nil }
-
-func (e *capturingExperimentExporter) firstGeneration() *sigilv1.Generation {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	if len(e.requests) == 0 || len(e.requests[0].GetGenerations()) == 0 {
-		return nil
-	}
-	return e.requests[0].GetGenerations()[0]
-}
-
-func (e *capturingExperimentExporter) hasGeneration(generationID string) bool {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	for _, request := range e.requests {
-		for _, generation := range request.GetGenerations() {
-			if generation.GetId() == generationID {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func (e *capturingExperimentExporter) generationCount(generationID string) int {
-	e.mu.Lock()
-	defer e.mu.Unlock()
-	count := 0
-	for _, request := range e.requests {
-		for _, generation := range request.GetGenerations() {
-			if generation.GetId() == generationID {
-				count++
-			}
-		}
-	}
-	return count
-}
-
-func TestRecordIOWithoutIOOrUsageDoesNotAttachGenerationID(t *testing.T) {
-	recorder := &experimentRecorder{}
-	recorder.push(http.StatusAccepted, map[string]any{"results": []map[string]any{{"score_id": "score-1", "accepted": true}}})
-	server := httptest.NewServer(recorder.handler(t))
-	defer server.Close()
-
-	client := newExperimentTestClient(t, server.URL)
-	trial := NewTrial(client, TrialRef{RunID: "run-empty-io", TestCaseID: "case-empty-io"})
-	trial.RecordIO(RecordIOOptions{
-		ModelProvider: "example",
-		ModelName:     "agent",
-		AgentName:     "support-agent",
-	})
-	score := trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-	if score.GenerationID != "" {
-		t.Fatalf("expected score without recorded IO or usage to omit generation_id, got %q", score.GenerationID)
-	}
-
-	accepted, err := trial.Flush(context.Background())
-	if err != nil {
-		t.Fatalf("flush trial: %v", err)
-	}
-	if accepted != 1 {
-		t.Fatalf("expected one accepted score, got %d", accepted)
-	}
-	if recorder.requestCount() != 1 {
-		t.Fatalf("expected only score export request, got %d", recorder.requestCount())
-	}
-	req := recorder.request(0)
-	scorePayload := req.Payload["scores"].([]any)[0].(map[string]any)
-	if _, ok := scorePayload["generation_id"]; ok {
-		t.Fatalf("score without recorded IO or usage must not send generation_id: %#v", scorePayload)
-	}
-}
-
-func TestBindGenerationWithRecordIOExportsBoundGeneration(t *testing.T) {
-	exporter := &capturingExperimentExporter{}
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		if req.URL.Path != "/api/v1/scores:export" {
-			http.NotFound(w, req)
-			return
-		}
-		if !exporter.hasGeneration("gen-recorded") {
-			http.Error(w, "recorded generation was not flushed before score export", http.StatusBadRequest)
-			return
-		}
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusAccepted)
-		_ = json.NewEncoder(w).Encode(map[string]any{
-			"results": []map[string]any{{"score_id": "score-1", "accepted": true}},
-		})
-	}))
-	defer server.Close()
-
-	client := NewClient(Config{
-		Tracer: noop.NewTracerProvider().Tracer("sigil-go-bound-record-io-test"),
-		GenerationExport: GenerationExportConfig{
-			Protocol:        GenerationExportProtocolHTTP,
-			Endpoint:        server.URL + "/api/v1/generations:export",
-			Auth:            AuthConfig{Mode: ExportAuthModeTenant, TenantID: "tenant-a"},
-			Insecure:        BoolPtr(true),
-			BatchSize:       10,
-			FlushInterval:   time.Hour,
-			QueueSize:       100,
-			MaxRetries:      0,
-			PayloadMaxBytes: 1 << 20,
-		},
-		API:                    APIConfig{Endpoint: server.URL},
-		testGenerationExporter: exporter,
-	})
-	t.Cleanup(func() { _ = client.Shutdown(context.Background()) })
-
-	trial := NewTrial(client, TrialRef{RunID: "run-recorded", TestCaseID: "case-recorded"})
-	trial.BindGeneration("gen-recorded", "conv-recorded")
-	trial.RecordIO(RecordIOOptions{Input: "question", Output: "answer", ModelProvider: "example", ModelName: "agent"})
-	trial.FinalScore(BoolScoreValue(true), ScoreOptions{})
-	accepted, err := trial.Flush(context.Background())
-	if err != nil {
-		t.Fatalf("flush trial: %v", err)
-	}
-	if accepted != 1 {
-		t.Fatalf("expected one accepted score, got %d", accepted)
-	}
-	generation := exporter.firstGeneration()
-	if generation == nil || generation.GetId() != "gen-recorded" || generation.GetConversationId() != "conv-recorded" {
-		t.Fatalf("expected recorded bound generation, got %#v", generation)
-	}
-}
-
-func TestTrialFlushFlushesBoundGenerationBeforeScores(t *testing.T) {
-	exporter := &capturingExperimentExporter{}
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		if req.URL.Path != "/api/v1/scores:export" {
-			http.NotFound(w, req)
-			return
-		}
-		if !exporter.hasGeneration("gen-bound") {
-			http.Error(w, "generation was not flushed before score export", http.StatusBadRequest)
-			return
-		}
-		w.Header().Set("Content-Type", "application/json")
-		w.WriteHeader(http.StatusAccepted)
-		_ = json.NewEncoder(w).Encode(map[string]any{
-			"results": []map[string]any{{"score_id": "score-1", "accepted": true}},
-		})
-	}))
-	defer server.Close()
-
-	client := NewClient(Config{
-		Tracer: noop.NewTracerProvider().Tracer("sigil-go-trial-flush-order-test"),
-		GenerationExport: GenerationExportConfig{
-			Protocol:        GenerationExportProtocolHTTP,
-			Endpoint:        server.URL + "/api/v1/generations:export",
-			Auth:            AuthConfig{Mode: ExportAuthModeTenant, TenantID: "tenant-a"},
-			Insecure:        BoolPtr(true),
-			BatchSize:       10,
-			FlushInterval:   time.Hour,
-			QueueSize:       100,
-			MaxRetries:      0,
-			PayloadMaxBytes: 1 << 20,
-		},
-		API:                    APIConfig{Endpoint: server.URL},
-		testGenerationExporter: exporter,
-	})
-	t.Cleanup(func() { _ = client.Shutdown(context.Background()) })
-
-	ctx, recorder := client.StartGeneration(context.Background(), GenerationStart{
-		ID:             "gen-bound",
-		ConversationID: "conv-bound",
-		Model:          ModelRef{Provider: "example", Name: "agent"},
-	})
-	recorder.SetResult(Generation{
-		ID:             "gen-bound",
-		ConversationID: "conv-bound",
-		Model:          ModelRef{Provider: "example", Name: "agent"},
-		Input:          []Message{UserTextMessage("question")},
-		Output:         []Message{AssistantTextMessage("answer")},
-	}, nil)
-	recorder.End()
-
-	trial := NewTrial(client, TrialRef{RunID: "run-bound", TestCaseID: "case-bound"})
-	trial.BindGeneration("gen-bound", "conv-bound")
-	trial.RecordIO(RecordIOOptions{Input: "question", Output: "answer", ModelProvider: "example", ModelName: "agent"})
-	trial.FinalScore(BoolScoreValue(true), ScoreOptions{Evaluator: &Evaluator{EvaluatorID: "exact", Version: "1", Kind: EvaluatorKindDeterministic}})
-	accepted, err := trial.Flush(ctx)
-	if err != nil {
-		t.Fatalf("flush trial: %v", err)
-	}
-	if accepted != 1 {
-		t.Fatalf("expected one accepted score, got %d", accepted)
-	}
-	if !exporter.hasGeneration("gen-bound") {
-		t.Fatal("expected bound generation to be flushed")
-	}
-	if got := exporter.generationCount("gen-bound"); got != 1 {
-		t.Fatalf("expected bound generation to be exported once, got %d", got)
-	}
-}
-
-func TestExperimentContextTagsExistingInstrumentationAndCapturesIDs(t *testing.T) {
-	exporter := &capturingExperimentExporter{}
-	client := NewClient(Config{
-		Tracer: noop.NewTracerProvider().Tracer("sigil-go-experiment-context-test"),
-		GenerationExport: GenerationExportConfig{
-			Protocol:        GenerationExportProtocolHTTP,
-			Endpoint:        "http://example.invalid/api/v1/generations:export",
-			Auth:            AuthConfig{Mode: ExportAuthModeTenant, TenantID: "tenant-a"},
-			Insecure:        BoolPtr(true),
-			BatchSize:       10,
-			FlushInterval:   time.Hour,
-			QueueSize:       100,
-			MaxRetries:      1,
-			InitialBackoff:  time.Millisecond,
-			MaxBackoff:      time.Millisecond,
-			PayloadMaxBytes: 1 << 20,
-		},
-		testGenerationExporter: exporter,
-	})
-	t.Cleanup(func() { _ = client.Shutdown(context.Background()) })
-
-	candidate := &Candidate{AgentName: "support-bot"}
-	run := NewExperimentRun(ExperimentOptions{Client: client, RunID: "run_existing", Name: "existing", Candidate: candidate})
-	ctx := WithConversationID(context.Background(), "ctx-conv")
-	ctx = WithAgentName(ctx, "ctx-agent")
-	ctx = run.Context(ctx)
-
-	ctx, recorder := client.StartGeneration(ctx, GenerationStart{
-		Model: ModelRef{Provider: "openai", Name: "gpt-5"},
-	})
-	recorder.SetResult(Generation{
-		Input:  []Message{UserTextMessage("hello")},
-		Output: []Message{AssistantTextMessage("world")},
-	}, nil)
-	recorder.End()
-	if err := client.Flush(ctx); err != nil {
-		t.Fatalf("flush: %v", err)
-	}
-
-	generation := exporter.firstGeneration()
-	if generation == nil {
-		t.Fatal("expected exported generation")
-	}
-	if generation.GetTags()[ExperimentRunIDTag] != "run_existing" {
-		t.Fatalf("expected experiment run tag, got %#v", generation.GetTags())
-	}
-	if generation.GetMetadata().GetFields()[ExperimentRunIDMetadataKey].GetStringValue() != "run_existing" {
-		t.Fatalf("expected experiment run metadata, got %#v", generation.GetMetadata())
-	}
-	if got := generation.GetConversationId(); got != "ctx-conv" {
-		t.Fatalf("expected context conversation id to win, got %q", got)
-	}
-	if got := generation.GetAgentName(); got != "ctx-agent" {
-		t.Fatalf("expected context agent name to win, got %q", got)
-	}
-	if ids := run.ProducedGenerationIDs(); len(ids) != 1 || ids[0] != generation.GetId() {
-		t.Fatalf("unexpected captured ids: %#v", ids)
 	}
 }
